Make browsing a native tool
