--- Extensions/Browsing/Manifest.json
@@ -1,23 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "com.nativ.browsing",
-  "version": "1.0.0",
-  "minimumNativVersion": "0.2.2",
-  "displayName": "Browsing",
-  "summary": "Fast, compact web search for tool-capable models.",
-  "developer": "Nativ",
-  "systemImage": "globe",
-  "included": true,
-  "enabledByDefault": true,
-  "runtime": "builtIn",
-  "extensionPoint": "com.nativ.extension",
-  "contributions": {
-    "sidebar": [],
-    "commands": [],
-    "shortcuts": [],
-    "settings": []
-  },
-  "permissions": [
-    "storage.namespaced"
-  ]
-}
