+++ Sources/Nativ/ControlPanelView.swift
@@ -1939,7 +1939,6 @@ struct ControlPanelView: View {
                 model: model,
                 chat: chat,
                 mcpHost: mcpHost,
-                extensionManager: extensionManager,
                 showsConfiguration: $isModelConfigurationVisible,
                 conversationWidthReduction: isFullScreen
                     ? 0
