+++ Sources/Nativ/NativSettings.swift
@@ -15,16 +15,13 @@ protocol ServerAPICredentialStoring {
 struct ServerAPIKeychain: ServerAPICredentialStoring {
     let service: String
     let account: String
-    let usesDataProtectionKeychain: Bool
 
     init(
         service: String = "dev.local.Nativ.server-api-key",
-        account: String = "nativ-server",
-        usesDataProtectionKeychain: Bool = true
+        account: String = "nativ-server"
     ) {
         self.service = service
         self.account = account
-        self.usesDataProtectionKeychain = usesDataProtectionKeychain
     }
 
     func load() throws -> String? {
@@ -80,15 +77,12 @@ struct ServerAPIKeychain: ServerAPICredentialStoring {
     }
 
     private var baseQuery: [String: Any] {
-        var query: [String: Any] = [
+        [
             kSecClass as String: kSecClassGenericPassword,
             kSecAttrService as String: service,
-            kSecAttrAccount as String: account
+            kSecAttrAccount as String: account,
+            kSecUseDataProtectionKeychain as String: true
         ]
-        if usesDataProtectionKeychain {
-            query[kSecUseDataProtectionKeychain as String] = true
-        }
-        return query
     }
 }
