+++ project.yml
@@ -100,9 +100,6 @@ targets:
         buildPhase: resources
       - path: Extensions/VoiceDictation/Manifest.json
         buildPhase: resources
-      - path: Extensions/Browsing
-        type: folder
-        buildPhase: resources
     dependencies:
       - target: NativServerKit
       - target: NativExtensionSDK
@@ -178,6 +175,8 @@ targets:
       - path: Sources/Nativ/Features/Chat/Tools/ChatListModelsTool.swift
       - path: Sources/Nativ/Features/Chat/Tools/ChatServerStatsTool.swift
       - path: Sources/Nativ/Features/Chat/Tools/ChatSwitchModelTool.swift
+      - path: Sources/Nativ/Features/Chat/Tools/BrowsingSearchTool.swift
+      - path: Sources/Nativ/Features/Browsing/BrowsingProvider.swift
       - path: Sources/Nativ/Features/Artifacts/Artifact.swift
       - path: Sources/Nativ/Features/Chat/ChatSessionStore.swift
       - path: Sources/Nativ/Features/ImageGeneration/ImageGenerationViewModel.swift
